Move test to separate file to reflect location of function under test
